feat: use api toolkit for user crud

+++ app/Http/Controllers/API/RoleController.php
@@ -3,11 +3,12 @@
 namespace App\Http\Controllers\API;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Role\UpdateRoleRequest;
 use App\Http\Requests\Role\CreateRoleRequest;
+use App\Http\Requests\Role\UpdateRoleRequest;
 use App\Http\Resources\Role\RoleResource;
 use App\Models\Role;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 
 class RoleController extends Controller
@@ -17,18 +18,26 @@ class RoleController extends Controller
 
     }
 
-    public function index(): AnonymousResourceCollection
+    public function index(Request $request): AnonymousResourceCollection
     {
-        $roles = Role::useFilters()->dynamicPaginate();
+        $request->validate([
+            'pagination' => 'in:none',
+            'per_page' => 'integer|min:1',
+            'search' => 'string',
+        ]);
+
+        $roles = Role::with('permissions')->useFilters()->dynamicPaginate();
 
         return RoleResource::collection($roles);
     }
 
     public function store(CreateRoleRequest $request): JsonResponse
     {
-        $role = Role::create($request->validated());
+        $validated = $request->validated();
+        $role = Role::create($validated);
+        $role->givePermissionTo($validated['permissions']);
 
-        return $this->responseCreated('Role created successfully', new RoleResource($role));
+        return $this->responseCreated('Role created successfully', new RoleResource($role->load('permissions')));
     }
 
     public function show(Role $role): JsonResponse
@@ -38,17 +47,33 @@ class RoleController extends Controller
 
     public function update(UpdateRoleRequest $request, Role $role): JsonResponse
     {
-        $role->update($request->validated());
+        $validated = $request->validated();
+        $role->update($validated);
+        $role->syncPermissions($validated['permissions']);
 
-        return $this->responseSuccess('Role updated Successfully', new RoleResource($role));
+        return $this->responseSuccess('Role updated Successfully', new RoleResource($role->load('permissions')));
     }
 
     public function destroy(Role $role): JsonResponse
     {
+        $role->load('permissions', 'users');
+
+        if ($role->permissions()->count() > 0) {
+            return $this->responseConflictError(
+                'Cannot remove role. Role has associated permissions.',
+                'Please first remove role\'s permissions',
+            );
+        }
+
+        if ($role->users()->count() > 0) {
+            return $this->responseConflictError(
+                'Cannot remove role. Role is assigned to one or more users.',
+                'Cannot remove role. Role is assigned to one or more users.',
+            );
+        }
+
         $role->delete();
 
         return $this->responseDeleted();
     }
-
-   
 }
